+++ examples/LinearSolver/updrhs.cl
@@ -1,1 +1,1 @@
-RhsTab[ cnt ] += (double4){.1,.2,.3,.4};
+RhsTab[ cnt ] += (float4){.1,.2,.3,.4};
